Fix issue
There is currently a bug where the last entity in the world doesn't save it's model correctly.

No idea where this behaviour is coming from, likely has to do with loading of the string rather then saving.

string saving should be switched to java's modified UTF 8 to allow for external programs to open .display files.